Update best_answer_id on questions table when answer removed in database layer

+++ database/migrations/2019_11_07_164806_create_answers_table.php
@@ -15,8 +15,8 @@ class CreateAnswersTable extends Migration
     {
         Schema::create('answers', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('question_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('question_id');
+            $table->unsignedBigInteger('user_id');
             $table->text('body');
             $table->Integer('votes_count')->default(0);
             $table->timestamps();
